Added error code framework
Each function which returns a value together with an error code can use
the ERROR_TYPE() types defined in this header. The types actually
represent a structure which holds the return value of the function and
the error code of the function.
If a function wants to return something else than a primitive type it
can define a new error type with the ERROR_TYPEDEF macro.