Complete rework of the VFS implementation
The VFS now only handles with a tree of vfs_node_t structures. The
filesystem functionalities are implemented elsewhere.
A new type of filesystem is introduced: the virtual filesystem. A
virtual filesystem is a filesystem of which the ressources are not
backed by an actual device.
This rework also introduced the new functionality of filesystem modules
which are virtual filesystem which are static (e.g. devfs). The
ressources are always the same independent of the mount point by which
they are accessed.
The cdifs filesystem driver is a wrapper to the filesystems which are
using cdi.
The devfs is a filesystem module by which each device can be accessed.
Also the mounting interface changed so that the caller now can specify a
filesystem.