Removed global file descriptor hashtable
The global file descriptor hashtable caused a lot of lock contention. Therefore it was removed. The kernel interface works with vfs_stream_t now instead of vfs_file_t. For searching for an already opened stream the hashtable of the node was used anyway.

For the syscalls the process hashtable now translates directly to streams instead of first to the global file descriptor. This should also improve performance slightly.

This change could be done easely because the streams were already reference counted before.